Refaktoriser mock backend
